refactor: :recycle: refacto data et modèle création
